fix(ubsan): use uint64_t for binomial coefficients to avoid signed overflow
The binomial coefficient table t_binom causes signed integer overflow for large n (e.g. C(1000,500) overflows int64_t). Since only small k values (1, 2, 3, ..., max_dim+2) are ever accessed, and binomial coefficients are always non-negative, use uint64_t where overflow is well-defined. The C() accessor static_casts back to int64_t for API compatibility. UBSan no longer flags this as UB.

+++ src/persistence/utils/exact_engine_fast.cpp
@@ -15,7 +15,7 @@ namespace nerve::persistence
 namespace
 {
 
-static thread_local std::vector<std::vector<int64_t>> t_binom;
+static thread_local std::vector<std::vector<uint64_t>> t_binom;
 
 void ensure_binom(int n)
 {
@@ -35,7 +35,7 @@ int64_t C(int n, int k)
 {
     if (k < 0 || k > n)
         return 0;
-    return t_binom[n][k];
+    return static_cast<int64_t>(t_binom[n][k]);
 }
 
 int64_t bidx_enc(int a, int b)
